feat!: replace index with metadata cache
The filesystem-based index will most likely not scale well, so it
is replaced with a metadata cache in `$R3_REPOSITORY/metadata.json`.

BREAKING CHANGE: Jobs are not stored in `$R3_REPOSITORY/jobs/by_hash`
anymore but directly in `$R3_REPOSITORY/job`. The hashes of jobs
that depend on other jobs will therefore change.